requirements.txt
streamlit>=1.28.0
openai>=1.0.0
sounddevice>=0.4.6
wavio>=0.0.7
python-dotenv>=1.0.0
numpy>=1.23.5
